Show tab color on the first tab after it's expanded (Fixes #1216)
Root issue here is that the NSWindowTab's accessoryView isn't added as a subview until the tab is actually shown. Since the tab bar isn't (usually) visible during Sequel Ace's first connection, our accessoryView didn't have a superview until then.

This causes two issues:
 1. The  `tabAccessoryColorView` didn't `wantsLayer`, so it was nil (evidently because it wasn't actually in the view hierarchy until the tab bar became visible) and was skipped in the "tabAccessoryColorView.layer?.backgroundColor =" assignment. Fixed by setting wantsLayer on the color view.
 2. The superview was nil, so the tabAccessoryView.snp.makeConstraints call to right-size the tab accessory view never ran (and if it had, it would have blown up because superview is nil). Fixed by moving the tabAccessoryView out into its own view that can listen for `viewDidMoveToSuperview` and apply the correct constraint there.
