+++ Source/Controllers/Window/SPWindowController.swift
@@ -43,37 +43,10 @@ import SnapKit
         }
 
         setupAppearance()
-        setupConstraints()
     }
 
     // MARK: - Accessory
-
-    private lazy var tabAccessoryView: NSView = NSView()
-    private lazy var tabAccessoryColorView: NSView = NSView()
-    private lazy var tabText: NSTextField = {
-        let text = NSTextField()
-        text.userActivity = .none
-        text.backgroundColor = .clear
-        text.isEditable = false
-        text.isHidden = false
-        text.alignment = .center
-        text.isBordered = false
-        text.textColor = .labelColor
-        return text
-    }()
-
-    private lazy var tabAccessoryViewImage: NSImageView = {
-        var image = NSImage(imageLiteralResourceName: "fallback_lock.fill")
-        if #available(macOS 11, *), let systemImage = NSImage(systemSymbolName: "lock.fill", accessibilityDescription: nil) {
-            image = systemImage
-        }
-        let imageView = NSImageView(image: image)
-        imageView.toolTip = NSLocalizedString("SSH Connected", comment: "Tooltip information text")
-        imageView.isHidden = true
-        return imageView
-    }()
-
-    private var tabAccessoryConstraintsSetup: Bool = false
+    private lazy var tabAccessoryView: SPWindowTabAccessory = SPWindowTabAccessory()
 }
 
 // MARK: - Private API
@@ -85,28 +58,8 @@ private extension SPWindowController {
         window?.contentView?.addSubview(databaseDocument.databaseView())
         databaseDocument.databaseView()?.frame = window?.contentView?.frame ?? NSRect(x: 0, y: 0, width: 800, height: 400)
 
-        tabAccessoryView.addSubviews(tabAccessoryColorView, tabAccessoryViewImage, tabText)
-
         if #available(macOS 10.13, *) {
             window?.tab.accessoryView = tabAccessoryView
-        }
-    }
-
-    func setupConstraints() {
-        tabAccessoryColorView.snp.makeConstraints {
-            $0.height.equalTo(5)
-            $0.bottom.leading.trailing.equalToSuperview()
-        }
-        tabText.snp.makeConstraints {
-            $0.bottom.equalTo(tabAccessoryColorView.snp.top)
-            $0.leading.top.equalToSuperview()
-            $0.trailing.equalTo(tabAccessoryViewImage.snp.leading)
-        }
-
-        tabAccessoryViewImage.snp.makeConstraints {
-            $0.size.equalTo(20)
-            $0.trailing.equalToSuperview()
-            $0.centerY.equalToSuperview()
         }
     }
 }
@@ -119,24 +72,12 @@ private extension SPWindowController {
         if #available(macOS 10.13, *) {
             window?.tab.title = tabTitle
         }
-        if tabAccessoryView.superview != nil {
-            tabText.stringValue = tabTitle
-        }
+        
+        tabAccessoryView.setTitle(title: tabTitle)
     }
 
     func updateWindowAccessory(color: NSColor?, isSSL: Bool) {
-        tabAccessoryColorView.layer?.backgroundColor = color?.cgColor
-        tabAccessoryViewImage.isHidden = !isSSL
-
-        if tabAccessoryView.superview != nil, !tabAccessoryConstraintsSetup {
-            tabAccessoryConstraintsSetup = true
-            tabAccessoryView.snp.makeConstraints {
-                $0.leading.equalToSuperview().offset(35)
-                $0.trailing.equalToSuperview().offset(-35)
-                $0.top.equalToSuperview().offset(5)
-                $0.bottom.equalToSuperview()
-            }
-        }
+        tabAccessoryView.update(color: color, isSSL: isSSL)
     }
 }
